# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TEQUALY Field Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contratos:
    get:
      operationId: listContratos
      tags: [contratos]
      summary: List contracts
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contrato"
    post:
      operationId: createContrato
      tags: [contratos]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContratoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrato"
  /contratos/{id}:
    get:
      operationId: getContrato
      tags: [contratos]
      summary: Get contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrato"
    patch:
      operationId: updateContrato
      tags: [contratos]
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContratoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrato"
    delete:
      operationId: deleteContrato
      tags: [contratos]
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Frentes ─────────────────────────────────────────────
  /frentes:
    get:
      operationId: listFrentes
      tags: [frentes]
      summary: List work fronts
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Frente"
    post:
      operationId: createFrente
      tags: [frentes]
      summary: Create a work front
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrenteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frente"
  /frentes/{id}:
    patch:
      operationId: updateFrente
      tags: [frentes]
      summary: Update work front
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrenteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frente"
    delete:
      operationId: deleteFrente
      tags: [frentes]
      summary: Delete work front
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Ativos ──────────────────────────────────────────────
...
  /conciliacoes/{id}/itens:
    get:
      operationId: listConciliacaoItens
      tags: [conciliacoes]
      summary: List line items extracted from documents imported into a reconciliation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConciliacaoItem"

  # ─── Exceções ─────────────────────────────────────────────
...
components:
  schemas:
...
    Contrato:
      type: object
      required: [id, numero, obra_id, status, created_at]
      properties:
        id:
          type: integer
        numero:
          type: string
        descricao:
          type: ["string", "null"]
        obra_id:
          type: integer
        valor:
          type: ["number", "null"]
        status:
          type: string
        data_inicio:
          type: ["string", "null"]
        data_fim:
          type: ["string", "null"]
        contratante:
          type: ["string", "null"]
        contratado:
          type: ["string", "null"]
        created_at:
          type: string
    ContratoInput:
      type: object
      required: [numero, obra_id, status]
      properties:
        numero:
          type: string
        descricao:
          type: string
        obra_id:
          type: integer
        valor:
          type: number
        status:
          type: string
        data_inicio:
          type: string
        data_fim:
          type: string
        contratante:
          type: string
        contratado:
          type: string
    ContratoUpdate:
      type: object
      properties:
        numero:
          type: string
        descricao:
          type: string
        valor:
          type: number
        status:
          type: string
        data_inicio:
          type: string
        data_fim:
          type: string
        contratante:
          type: string
        contratado:
          type: string
    Frente:
      type: object
      required: [id, nome, obra_id, status, created_at]
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        obra_id:
          type: integer
        responsavel:
          type: ["string", "null"]
        status:
          type: string
        created_at:
          type: string
    FrenteInput:
      type: object
      required: [nome, obra_id, status]
      properties:
        nome:
          type: string
        descricao:
          type: string
        obra_id:
          type: integer
        responsavel:
          type: string
        status:
          type: string
    FrenteUpdate:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        responsavel:
          type: string
        status:
          type: string
...
    ConciliacaoItem:
      type: object
      required: [id, conciliacao_id, origem_arquivo, formato_origem, extracao_confianca, created_at]
      properties:
        id:
          type: integer
        conciliacao_id:
          type: integer
        codigo:
          type: ["string", "null"]
        descricao:
          type: ["string", "null"]
        unidade:
          type: ["string", "null"]
        qtd_enviado:
          type: ["string", "null"]
        qtd_faturado:
          type: ["string", "null"]
        qtd_devolvido:
          type: ["string", "null"]
        qtd_em_campo:
          type: ["string", "null"]
        qtd_indenizado:
          type: ["string", "null"]
        valor_unitario:
          type: ["string", "null"]
        origem_arquivo:
          type: string
        formato_origem:
          type: string
        extracao_confianca:
          type: string
        created_at:
          type: string
...